+++ lib/grape/validations/types.rb
@@ -24,7 +24,12 @@ module Grape
     module Types
       # Instances of this class may be used as tokens to denote that
       # a parameter value could not be coerced.
-      class InvalidValue; end
+      class InvalidValue
+        attr_reader :message
+        def initialize(message = nil)
+          @message = message
+        end
+      end
 
       # Types representing a single value, which are coerced through Virtus
       # or special logic in Grape.